+++ tenpy/models/xxz_chain.py
@@ -60,10 +60,10 @@ class XXZChain(CouplingModel, NearestNeighborModel, MPOModel):
         #     from tenpy.networks.site import SpinHalfSite
         #     site = SpinHalfSite(conserve='Sz')
         # 4) lattice
-        lat = Chain(L, site, bc_MPS=bc_MPS)
-        bc_coupling = 'periodic' if bc_MPS == 'infinite' else 'open'
+        bc = 'periodic' if bc_MPS == 'infinite' else 'open'
+        lat = Chain(L, site, bc=bc, bc_MPS=bc_MPS)
         # 5) initialize CouplingModel
-        CouplingModel.__init__(self, lat, bc_coupling)
+        CouplingModel.__init__(self, lat)
         # 6) add terms of the Hamiltonian
         # (u is always 0 as we have only one site in the unit cell)
         self.add_onsite(-np.asarray(hz), 0, 'Sz')