+++ CHANGELOG.rst
@@ -9,8 +9,10 @@ The project adheres `semantic versioning <http://semver.org/spec/v2.0.0.html>`_
 
 Backwards incompatible changes
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-- The argument order of :meth:`Lattice` could be a tuple ``(priority, snake_winding)`` before.
-  This is no longer valid and needs to be replaced by ``("standard", snake_winding, priority)``
+- The argument order of :class:`tenpy.models.lattice.Lattice` could be a tuple ``(priority, snake_winding)`` before. 
+  This is no longer valid and needs to be replaced by ``("standard", snake_winding, priority)``.
+- Moved the boundary conditions `bc_coupling` from the :class:`tenpy.models.model.CouplingModel` into the :class:`tenpy.models.lattice.Lattice` (as `bc`).
+  Using the parameter `bc_coupling` will raise a FutureWarning, one should set the boundary conditions directly in the lattice.
 
 Added
 ^^^^^